Update CMake configuration and refactor main.cpp
The CMake configuration and build details have been updated to use a newer version of CMake and the associated tools. In the main.cpp file, the output file format has been modified to better represent the stixel information. Image paths were also updated to point to the new testing dataset.